add the ability to perform AIMD using BAGEL or ORCA

+++ script/wrapbagel.sh
@@ -5,29 +5,35 @@ usage() {
 Usage: $(basename $0) [options]
 
 Options:
-  -s <system> System file. (default: ${METHOD})
+  -a <active> Number of active orbitals. (default: ${NACTIVE})
   -b <basis>  Basis set used to specify the atomic orbitals. (default: ${BASIS})
-  -c <charge> Charge of the system. (default: ${CHARGE})
-  -p <nspin>  The number associated with the spin states: 0 for singlet, 1 for doublet, 2 for triplet, etc. (default: ${NSPIN})
+  -c <closed> Number of closed orbitals. (default: ${NCLOSED})
+  -g <charge> Charge of the system. (default: ${CHARGE})
+  -k <nstate> Number of states to compute. (default: ${NSTATE})
   -m <method> Method to perform. (default: ${METHOD})
+  -p <nspin>  The number associated with the spin states: 0 for singlet, 1 for doublet, 2 for triplet, etc. (default: ${NSPIN})
+  -r <retain> Retain the .bagel directory and create .all files. (default: ${RETAIN})
+  -s <system> System file. (default: ${METHOD})
   -h          Display this help message and exit.
 EOF
 
-}; SYSTEM="molecule.xyz"; BASIS="svp"; CHARGE=0; NSPIN=0; METHOD="sa2-casscf(2,4)"
+}; SYSTEM="molecule.xyz"; BASIS="svp"; CHARGE=0; NSPIN=0; METHOD="casscf"; NSTATE=2; NACTIVE=2; NCLOSED=4; RETAIN=0
 
-while getopts "b:c:p:m:s:h" OPT; do case "$OPT" in
+while getopts "a:b:c:g:k:m:p:rs:h" OPT; do case "$OPT" in
+  a ) NACTIVE="$OPTARG" ;;
   b ) BASIS="$OPTARG" ;;
-  c ) CHARGE="$OPTARG" ;;
+  c ) NCLOSED="$OPTARG" ;;
+  g ) CHARGE="$OPTARG" ;;
+  k ) NSTATE="$OPTARG" ;;
   m ) METHOD="$OPTARG" ;;
   p ) NSPIN="$OPTARG" ;;
+  r ) RETAIN=1 ;;
   s ) SYSTEM="$OPTARG" ;;
   h ) usage && exit 0 ;;
   \? ) usage && exit 1 ;;
 esac done
 
-mkdir -p .bagel; METHOD=${METHOD^^}; NATOM=$(awk 'END {print NR - 2}' $SYSTEM); NSTATE=$(sed -En 's/^SA([0-9]+)-CASSCF.*$/\1/p' <<< $METHOD)
-
-NACTIVE=$(sed -En 's/^SA[0-9]+-CASSCF\(([0-9]+),.*$/\1/p' <<< $METHOD); NCLOSED=$(sed -En 's/^SA[0-9]+-CASSCF\([0-9]+,([0-9]+)\).*$/\1/p' <<< $METHOD)
+mkdir -p .bagel; METHOD=${METHOD^^}; NATOM=$(awk 'END {print NR - 2}' $SYSTEM)
 
 cat << EOT > .bagel/bagel.json
 { "bagel" : [
@@ -40,10 +46,18 @@ cat << EOT > .bagel/bagel.json
   "geometry" : [
 EOT
 
-awk 'NR == 1 {A = $1} NR > 2 {printf("{\"atom\" : \"%s\", \"xyz\" : [%20.14f, %20.14f, %20.14f]}%s\n", $1, $2, $3, $4, NR == A + 2 ? "" : ",")}' molecule.xyz >> .bagel/bagel.json
+awk 'NR == 1 {A = $1} NR > 2 {printf("{\"atom\" : \"%s\", \"xyz\" : [%20.14f, %20.14f, %20.14f]}%s\n", $1, $2, $3, $4, NR == A + 2 ? "" : ",")}' $SYSTEM >> .bagel/bagel.json
 
 cat << EOT >> .bagel/bagel.json
   ]
+},
+EOT
+
+[[ -f .bagel/orbitals.archive ]] && cat << EOT >> .bagel/bagel.json
+{
+  "title" : "load_ref",
+  "file" : "orbitals",
+  "continue_geom" : false
 },
 EOT
 
@@ -54,7 +68,7 @@ cat << EOT >> .bagel/bagel.json
 EOT
 
 for I in $(seq 0 $((NSTATE - 1))); do
-    awk -v I=$I 'BEGIN {printf("{\"title\" : \"force\", \"target\" : %d},", I)}' >> .bagel/bagel.json
+    awk -v I=$I -v NSTATE=$NSTATE 'BEGIN {printf("{\"title\" : \"force\", \"target\" : %d}%s", I, NSTATE > 1 ? "," : "")}' >> .bagel/bagel.json
 done
 
 for I in $(seq 0 $((NSTATE - 1))); do
@@ -72,28 +86,42 @@ cat << EOT >> .bagel/bagel.json
     "nspin" : $NSPIN,
     "nact" : $NACTIVE,
     "nclosed" : $NCLOSED,
-    "nstate" : $NSTATE
+    "nstate" : $NSTATE,
+    "maxiter" : 2000
   }]
+},
+{
+  "title" : "save_ref",
+  "file" : "orbitals"
 }
 ]}
 EOT
 
-jq . .bagel/bagel.json > .bagel/bagel.json.tmp && mv .bagel/bagel.json.tmp .bagel/bagel.json
+jq . .bagel/bagel.json > .bagel/bagel.json.tmp && mv .bagel/bagel.json.tmp .bagel/bagel.json && rm -f ENERGY.mat GRADIENT.mat NACV.mat
 
-cd .bagel && BAGEL bagel.json ; cd ..
+cd .bagel && BAGEL bagel.json | tee bagel.out ; cd ..
 
-awk -v NATOM=$NATOM -v NSTATE=$NSTATE 'BEGIN {print NSTATE + 3 * NATOM * NSTATE + 3 * NATOM * (NSTATE * NSTATE - NSTATE) / 2}' > ENGRADNACV.mat
+if grep -q "EXCEPTION RAISED"             .bagel/bagel.out; then exit 1; fi
+if grep -q "Max iteration reached in SCF" .bagel/bagel.out; then exit 1; fi
 
-awk '{printf("%20.14f\n", $1)}' .bagel/ENERGY.out >> ENGRADNACV.mat
+awk -v NATOM=$NATOM -v NSTATE=$NSTATE 'BEGIN {print NSTATE, 1}'                                     >   ENERGY.mat
+awk -v NATOM=$NATOM -v NSTATE=$NSTATE 'BEGIN {print 3 * NATOM * NSTATE, 1}'                         > GRADIENT.mat
+awk -v NATOM=$NATOM -v NSTATE=$NSTATE 'BEGIN {print 3 * NATOM * (NSTATE * NSTATE - NSTATE) / 2, 1}' >     NACV.mat
+
+awk '{printf("%20.14f\n", $1)}' .bagel/ENERGY.out >> ENERGY.mat
 
 for I in $(seq 0 $((NSTATE - 1))); do
-    awk 'NF != 0 && NR > 1 {printf("%20.14f\n%20.14f\n%20.14f\n", $2, $3, $4)}' .bagel/FORCE_$I.out >> ENGRADNACV.mat
+    awk 'NF != 0 && NR > 1 {printf("%20.14f\n%20.14f\n%20.14f\n", $2, $3, $4)}' .bagel/FORCE_$I.out >> GRADIENT.mat
 done
 
 for I in $(seq 0 $((NSTATE - 1))); do
     for J in $(seq $((I + 1)) $((NSTATE - 1))); do
-        awk 'NF != 0 && NR > 1 {printf("%20.14f\n%20.14f\n%20.14f\n", $2, $3, $4)}' .bagel/NACME_${I}_$J.out >> ENGRADNACV.mat
+        awk 'NF != 0 && NR > 1 {printf("%20.14f\n%20.14f\n%20.14f\n", $2, $3, $4)}' .bagel/NACME_${I}_$J.out >> NACV.mat
     done
 done
 
-rm -rf .bagel
+if [[ $RETAIN -eq 1 ]]; then
+    cat $SYSTEM >> $SYSTEM.all && cat .bagel/bagel.out >> bagel.out.all
+else
+    rm -rf .bagel
+fi
